formatting: Added .clang-format for consistant formatting
clang-format is a popular tool to keep c/c++ code formatting consistant.
I mostly guessed a config, based on the coding style already used, but
there are probably improvements to be made to it. This is mostly to
ensure consistancy accross the whole code, and to make it easier for
other people contributing to have properly formatted code.

To format the code, once you have clang-format installed, run the
following command:

 $ clang-format -i cwordle.c

Some IDE/Text Editors support clang-format by default, or have
extensions to make it easy to reformat the code
This should be ran by everyone before comitting anything.

List of all options:
https://clang.llvm.org/docs/ClangFormatStyleOptions.html